authorize user controller

+++ FlowPanelApp/Controllers/UserController.cs
@@ -1,41 +1,71 @@
 ﻿using FlowPanelApp.Models;
 using FlowPanelApp.Services.UserService;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace FlowPanelApp.Controllers
 {
     public class UserController : Controller
     {
-        private readonly IUserService _userService;
+        private readonly IUserService _userService;       
         public UserController(IUserService userService)
         {
-            _userService = userService;
+            _userService = userService;                 
         }
         [Authorize]
         public async Task<IActionResult> Index()
         {
-            var model = await _userService.GetAll();
-            return View(model);
+            if (HttpContext.User.Identities.Select(x => x.RoleClaimType == "Admin").FirstOrDefault())
+            {
+                var model = await _userService.GetAll();
+                return View(model);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         [Authorize]
         public IActionResult CreateUser()
         {
-            return View();
+            if (HttpContext.User.Identities.Select(x => x.RoleClaimType == "Admin").FirstOrDefault())
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         [Authorize]
         public async Task<IActionResult> Create(User user)
         {
-            await _userService.AddUser(user);
-            return RedirectToAction("Index");
+            if (HttpContext.User.Identities.Select(x => x.RoleClaimType == "Admin").FirstOrDefault())
+            {
+                await _userService.AddUser(user);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         [Authorize]
         public async Task<IActionResult> Delete(long userId)
         {
-            var user = await _userService.GetUserById(userId);
-            await _userService.DeleteUser(user);
-            return RedirectToAction("Index");
+            if (HttpContext.User.Identities.Select(x => x.RoleClaimType == "Admin").FirstOrDefault())
+            {
+                var user = await _userService.GetUserById(userId);
+                await _userService.DeleteUser(user);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
     }
 }
